ref(KUI-1176): rename filteredData
